ADD laradock as a submodule
Instead of cloning a separate instance of laradock, add laradock as a
submodule. The submodule is our own fork of laradock, on the lti-shim
branch. No customization right now, but later on, we'll need to enable
php-gmp in laradock when we start using the JWT Framework.

I find the suggestion to not commit .env a bit unnecessary for our
development environment as the whole point of using docker is that we
have identical environments and there shouldn't be individual
customizations. So I've committed the .env for laradock. Not yet
committed the .env for the app though, we'll try this arrangement out
for a bit and see if there's anything I've missed.

Updated readme now that we don't need to clone laradock. Also added some
bits about things to be aware of for deployment.

Switched to using postgres for the database. Since phpmyadmin won't work
for postgres, switched to using adminer.