ADD lti component skeleton
This commit is for reference only. This is the start of an abstracted
approach to the LTI component that is designed from the start to be a
library. This is probably a premature optimization type of approach, so
have been abandoned for now. We'll do this without worrying about being
a library first.

The idea is to have an abstracted log and storage mechanism so it can be
used with any framework. A trivial log mechanism has been implemented as
an example.

Because the LTI component is outside of the `app/` directory, we need to
make the autoloader aware of it. Looks like there is a PHP standard for
autoloaders called PSR-4 that Laravel is using. We need to edit
`composer.json` to give the namespace prefix and the directory we want
to load.