ADD lti component skeleton
LTI specific code is located in the `lti/` directory.

Because the LTI component is outside of the `app/` directory, we need to
make the autoloader aware of it. Looks like there is a PHP standard for
autoloaders called PSR-4 that Laravel is using. We need to edit
`composer.json` to give the namespace prefix and the directory we want
to load.

In the `lti/` directory, the idea is to roughly organize the code
according to which LTI spec they're defined in. For now, we only have
the core spec, which I've defined in the subnamespace `Launch`. I didn't
want to call it `Core` as that can be unintentionally confusing, and
since documentation still calls it a 'launch', it seems to be the better
name.

Note, nothing is functional yet in this commit.