ADD launch auth request & response stages
Can now use the reference implemenation platform to launch to this tool.
The redirect_uri and the auth request form uri is hardcoded though and
so will work with my instance of the RI platform.

The auth request shows a page with form that the user has to click a
submit button on in order to proceed.

The auth response just shows a success page if the response passed
validation.

Laravel has a built in csrf token protection on all forms. We need to
disable this protection for the lti requests as the platform is doing a
csr request but we want to allow it. This is configured in the
VerifyCsrfToken middleware.