ADD rsa key storage and serve up jwks
Part of working on the LTI platform side of things is that we need to
generate a JWT that is signed by our own private key and verified by our
public key. So we need a way to store and retrieve RSA keys.

The KeyStorage solution in this commit just returns hardcoded values,
it will need to be replaced by a database backed solution later.

Serve up the public key as a JSON Web Key Set at the /lti/jwks url. The
LTI reference implementation doesn't actually support entering keys as
JWK, so we need to manually translate it to x509 PEM format. But Canvas
allows input as JWK and if it weren't being served on localhost, we
could actually give Canvas the url and it'll retrieve the key for us
automatically.

Removed example feature test that was broken since the web root no
longer serves a page. Apparently, the web root was only an issue because
it was using a closure/anonymous function, controllers should work fine.
Seems to be an issue that only happens with the deployed version, not
the laradock run one. I've moved the reference key generation function
into the JWKSController to avoid this issue.