ADD platform launch login stage
Can now send a login request to a Reference Implementation tool, url is
hardcoded.

Same as the ToolLaunch class, we now have a PlatformLaunch class for
checking and returning the parameters that needs to be sent during
launch.

Same as the ToolLaunchController, we now also have a
PlatformLaunchController that calls on PlatformLaunch to process
requests.

View blade templates have now been split up into tool and platform
sides. The pages are very similar right now and we might be able to
unify the perhaps later on, I'll know once more functionality has been
implemented. For now, splitting them makes the work easier.

Renamed test case to make it more clear that it's only checking incoming
parameters and not the parameters that are sent out.