EDIT PlatformLaunch enable auth req & resp stages
Can now receive auth request at /lti/launch/platform/auth and can now
send out auth response with a properly signed id_token.

I could not get the reference implementation tool to work. It's showing
a 500 server error even on stages that worked earlier, so it seems to be
having issues server side and not on my side. I switched to developing
against the LTI 1.3 PHP Example Tool and was able to get the platform
launch to work. The example tool didn't work out of the box for me
though, had to rewrite the docker files to get it to work, so it's kinda
customized.

Added library for RSA256 signature support, which required enabling the
php-gmp extension on laradock: web-token/jwt-signature-algorithm-rsa