ADD lti config to store the shim's own lti info
Information on other platforms and tools are stored in the database. We
need to have the same kind of information for the shim itself, such as
our 'iss' and launch url locations. This starts that by using Laravel's
config facilities.

TooLaunch has been modified to use these config values.

For now, all our config can be inferred by the app.url setting, so it's
not that complicated. But later on, we will need to add configurations
that can be overridden by .env files.