FIX param check tests, timestamp default cur time
More stricter param checking was implemented so the tests also needed to
be updated.

`iss` now needs to be stored in the database, so we're now taking
advantage of Laravel's data factory mechanism to create those for the
tests.

Fixed bug where a call to strpos() had the argument order the wrong way
around. This meant valid `target_link_uri` was being wrongly rejected.

Also now creating the 'updated_at' and 'created_at' table columns
manually. The Laravel default doesn't seem to enable database defaults,
so it relies on the ORM to provide values on insert/updates. I don't
like this since it means that manually inserted data would lack default
values. Creating the timestamps manually allows to at least specify that
it defaults to SQL's current_timestamp on create.

It looks like postgres doesn't have 'on update current_timestamp'
support, so I didn't try to force that default on it. You're supposed to
create a trigger to the same effect but I probably shouldn't spend time
on it right now.