EDIT ToolLaunch store params from login in state
Some params needs to be validated against the id_token, so we need to
store them.

The state param in the auth request is meant for Tool statekeeping. It
also serves a secondary function as CSRF protection in the original
OAuth spec, in a similar manner to the nonce param. This is perhaps why
the LTI spec has this contradiction where state is optional in the auth
request, but required in the auth response. The OpenID spec clearly
says state is only required in the auth response only if it was included
in the auth request. This means it's probably better if we treat it as
required in both stages in LTI. Since it's required, we might as well
put it to use storing the params that needs to be validated later.