EDIT ToolLaunch verifies state & id_token JWT
ToolLaunch now properly checks the signatures of state & id_token JWT.
It also checks required params of id_token.

Not sure yet if we should do a more thorough validation as the specs do
list more required claims that I didn't check.

RequestChecker is now the more general ParamChecker as I wanted to use
it with the JWT claims too.

Realized that that abort() function didn't record the LTIException stack
trace, so added a report() call to record them to the logs.