ADD Tool and Deployment tables, models, & seeder.
Deployments are each instance of deployment_id we've seen. The
deployment_id is needed for us to know what tool to connect to with each
LTI launch. Each instance of a tool link has a deployment_id in the
platform. Since we want instructors to be able to configure which link
goes to which tool, the shim will need to  use the deployment_id to identify which tool to go to.

A deployment_id is also platform specific, so a Deployment also has an
associated Platform.

Note that the target_link_uri is supposed to be configured beforehand
somehow. I have the feeling it should be something configurable by
instructors per deployment, but am not sure. For now, I've put it as a
Tool field.

Also added corresponding Tool seeder that'll create Tool and Deployment
entries in the database for dev.

Fixed namespace issues where I forgot to use the plural Models.