EDIT Tool/Platform Launch pass data by sessions
After the last stage auth response processing in the ToolLaunch, the
ToolLaunchController redirect to the first stage of the
PlatformLaunchController.

The Tool and Platform sides are now "talking" to each other through
sessions. At least the for the first stage OIDC login of the
PlatformLaunch.

We're also now retriving Tool and Deployment data from the database for
the OIDC login on the PlatformLaunch instead of using hardcoded values.

It turns out that Laravel sessions converts keys with periods in them
into a multi-level array. So if you ask it store ['a.b' => 'val'] it
would convert it to ['a' => ['b' => 'val']]. This was problematic for us
as id_token contains a lot of LTI URI being used as keys, which were
being split up when I put it into sessions. So I had to preserve the
original id_token as is when stored in sessions and deserialize it when
we need it.

Fixed lti_deployment_id being expected to be present when it's an
optional param.