FIX invalid JWT throws a 500 instead of a 400 err
When trying to deserialize a JWT, if the token isn't actually a JWT,
it'll throw a error that gets propagated without being wrapped into an
LTI error. Catching this error and the wrapping it in LTIException
causes the error to be properly treated as an error with the request,
not the server.