EDIT PlatformFactory create related PlatformClient
Platforms need at least one associated client, so we when we create one
using the PlatformFactory, it should also create an associated
PlatformClient.

Using the afterCreate callback on the factory since primary id isn't
given until it's written to the database. afterMake's callback means
the platform hasn't been written to the database yet so has no id.

Note that the PlatformClientFactory should not be used by itself, as it
will error out due to not having a platform_id.