ADD tool and platform rsa key tables
Each platform and tool might have multiple keys. I tried to find a way
to store all the keys in a single table but couldn't find an easy way to
maintain relations without dropping into raw sql. So as a simple
alternative, I've instead created two tables for storing keys, one each
for the platform and tool.

Note that in practice,  we should only have to store public keys for
signature verification.

Added a PlatformKeyFactory that generates a key on the fly. It should be
noted that factories are intended for use with test cases, so the
generated key that is stored contains BOTH public and private keys. This
makes it easier for the test cases as we need the private key to
generate signatures.

Modified the PlatformFactory to generate an associated key when a
platform is created.

The ToolKey table was created but since the LTI launch does not need it,
we don't need to use it yet. It will be needed when implement the other
specs.