EDIT rename 'public_key' to 'key' in RSA key table
Basically acknowledging that the 'key' column might store both the
public and private keys, so shouldn't be named 'public_key'. This is
mainly so that we can store the shim's own platform information in the
tables instead of using the config files. A secondary effect is that
we're more in line with the spec. So if there's some insane
implementation out there where they expect all keys to be generated by
the platform alone or the tool alone (allowed by the spec), we can still
work with them.

Since we don't want to accidentally leak the private key, both
PlatformKey and ToolKey models now has a public_key accessor that
makes sure only the public key part of the key is returned.

Also added an index on kid and platform/tool id pairs. This ensures that
kids must be unique within each platform/tool.