EDIT start store shim's platform & tool info in db
This started with realizing that it seems a bit weird we're building all
these tables for storing platform and tool information but not using
them to store the shim's own platform and tool config. So instead of
relying on the config, I'm trying to move the shim's own platform and
tool info into the db too.

The main thing is that there are now two new lti config:
* own_platform_id: the id of the shim's entry in the platforms table
* own_tool_id: the id of the shim's entry in the tools table

And we'll now be working on removing the rest of the lti config.

Each Platform and Tool model now also have a static function that can be
called to retrieve the shim's entry, e.g: Platform::getOwnPlatform()

Minor non-related changes: added some comments to some database tables.