ADD encryption keys table
In preparation for using encrypted JWT to keep state instead of cookies.

Create the encryption_keys table and add the associated seeder and
factory.

Since all the RSA key models share the same need to return a JWK object,
I've abstracted this out to the AbstractRsaKey model. So ToolKey,
PlatformKey, and EncryptionKey all extends this model.