ADD EncryptedState class to generate encrypted JWT
The state param in auth req & resp in the ToolLaunch was originally a
JWS. I've now replaced it with a JWE.

The original idea was to use the JWE to replace the data passing that
was being done with sessions between the Tool and Platform side. This
would allow us to go cookieless by removing the need for the built
inLaravel sessions. Unfortunately, due to the length of the JWE (~1000
characters right now and we'll probably need more) this might be an
issue with GET requests.  Most browsers shouldn't have issues with long
GET requests, but IE and Edge both have a rather low limit of 2k
characters and are still somewhat of a concern.

So I guess we'll have to go back to a more traditional database backed
sessions, with the JWE passing the session ID.