ADD LtiSession, now using cookieless sessions
Now no longer using Laravel's built-in sessions as those use cookies,
which we want to phase out for the LTI parts, to avoid issues with the
SameSite change. The new LtiSession model and associated table replaces
that.

Instead of a cookie storing the session id, we're passing along a JWE
storing the session id. This JWE is passed using the the
lti_message_hint parameter.

The new session currently stores all of the id_token's claims plus a
deployment_id that lets us look up the deployment in the database.

Note that the deployment table also contains a column named
deployment_id, this is confusing and should be renamed.

login_hint is also currently passed in the LTI session, but this
progbably will be changed once we implement filtering.

TODO: fix test cases.