ADD whitelist filtering of params
This starts to implement the filtering mechanisms that is the whole
point of the shim.

The filtering only needs to be applied on the shim's platform side, as
that is the one that talks to 3rd party tools.

Currently, PlatformLaunch has a list of filters that is applied to all
parameters that is sent out. Filters have to implement the
FilterInterface and should be placed in the Filters directory. The
filter function just takes in an array of params and return the filtered
array of params.  Hopefully this is sufficiently expandable for all
filtering tasks that we want to do.

The WhitelistFilter will remove any parameter names that it does not
recognize. Only parameters that are listed in the relevant specs should
be allowed through. It's pretty strict currently so we probably need to
adjust it in real use. I can see tools that use some functionality from
OAuth/JWT/OpenID that isn't explicitly mentioned in the LTI specs but
can be expected to be implemented due to LTI being a derivative spec.