ADD UserFilter to filter out user information
UserFilter replaces user data with fake data. This is on a
per-deployment basis, so a user's fake identity should be different on
each tool.

Currently giving out fake login_hint, name, and email. The 'sub' claim
in id_token is replaced with the fake login_hint.

The fake data is currently generated by the Faker library. I'm not sure
how good the randomness is, so we should probably test that it's fine
for use in prod.

Had to add the LtiSession as a parameter to filter. The params by itself
wasn't sufficient to properly implement the UserFilter. Luckily, all
stages of the PlatformLaunch has an LtiSession, so the change went
smoothly.

The non_lti_id field in the lti_users table is meant for puid in the UBC
use case. I'm thinking of making it configurable, so we can use it with
Canvas custom parameters, but it'll need to be implemented after we've
got this running with Canvas.

Note that if login_hint or sub isn't in sessions, the UserFilter assumes
there's no user data to filter out.

Note that roles are now no longer empty, they are passed as is from the
original platform.