UPDATE laradock & readme
Updated laradock to also start the LTI example tool, so we don't have to
start it separately. It should work out of the box with the seed data.

Readme updated with brief descriptions of the LTI components.

The root route is now working again as a quick way to show that the app
is working.