ADD platform login & auth req tests, FIX filter
Platform now has tests for all 3 stages of the LTI launch.,

The new tests revealed a regression where the filter wasn't properly
applied to the id_token. This was fixed.

Also did some additional minor fixes around required params in the
id_token. The params weren't specified in the LTI Security Spec, but
they are specified in the JWT specs, so I've added them just in case.

The LTI spec is problematic on whether 'state' is a required parameter
in the auth response. I've deferred to the OpenID Connect spec instead,
where 'state' is only required if it was given in the auth request.

Added short instructions on how to run the tests in readme.