ADD DepolymentFilter to send fake deployment ID
Passing the LTI deployment ID straight through might allow some personal
information to be inferred indirectly. So I've added a filter that'll
substitute the real LTI deployment ID with a fake one.

Added the fake_lti_deployment_id column to the deployments table.

Like LtiUser, we fill in the fake data with a fillFakeFields() method on
the Deployment model. I tried using an attribute accessor to auto
populate the fake_lti_deployment_id field, but it didn't work.

There's a strange error on save that seems to be complaining about the
ORM not properly treating the timestamps as a Postgres timezone enabled
timestamp. The fix was to explicitly tell the model what timestamp
format to use. It's very strange that the Deployment model has this
issue but the LtiUser model doesn't, but it's not causing an issue right
now, so I just hope it doesn't bite us later.

Also fixed the tests to check for this fake LTI deployment ID.