ADD ability to pull public keys from JWKS url
Needed to give the Platform and Tool models a parent class for accessing
the keys by key id. This is the AbstractLtiService class. The getKey()
method, if given a kid, will first look in the database to see if it can
find the key. If it can't, it'll try to pull new keys from the JWKS url
and then see if it can find the key again. If no kid was given, then
it'll just return the newest key for that Tool or Platform.

JwksUpdater only has one job, it pulls the the keys from the JWKS url
and adds it to the database. It currently has an issue where it'll
insert duplicate keys into the database as it doesn't check if the key
already exists.

Getting the kid from the token is unfortunately more involved than it
should be. The JWT Framework library's Easy components are focused on
verifying the JWT and doesn't give you access to the claims until after
verification is done. This creates the conundrum where we can't get the
kid until after verification but we need the kid to get the key for
verification. To work around this, we have to deserialize the token
ourselves to grab the kid before verification.

Not very happy with the exceptions I've set up around this, will
probably need to refactor them in use.