FIX multicolumn indexes are not unique & misc
I've been working under the misunderstanding that creating an index
means we're also enforcing uniqueness on that index. Turns out this was
wrong, it's only creating an index. We need to use unique() in order to
enforce uniqueness. This turned up because JwksUpdater was putting in
duplicate keys somehow even though I thought the table should've
rejected duplicate keys.

Small fixes:
* Moved save operation from UserFilter to LtiUser model.
* Brace formatting following php standard.
* target_link_uri changed to match the fact that the shim's tool side
redirects user to the platform side.