FIX JwksUpdater error when inserting duplicate key
With the unique constraint fix, the database now correctly complains
when we try to insert a duplicate key. So JwksUpdater will only work the
first time, then always fail all subsequent calls since there'll be
duplicate keys.

I couldn't get any of the ORM aware create/save methods to work. Seems
like there's no way to ignore the insertion errors using Laravel's ORM.
There are ways to do this for single insertions, but it seems rather
expensive, as it searches the database for an existing entry before
trying to create a new one. This might add up if we have a long list of
keys.

So we have to use the lower level query builder's insertOrIgnore()
method to keep the number of queries down. This means we have to
manually specify the foreign key column name and value, since the query
builder is not relations aware. Otherwise, it's still mostly the same
code.