EDIT tool launch pick target tool by custom claim
Instead of having the user pick what tool they want to go to on the
first launch, the target tool will instead be expected to be in the
id_token's custom claim.

This means we'll see a different workflow on the Canvas side. Instead of
having only a single LTI Shim external tool, there will be many of them,
each one configured to launch (via the shim) to a single tool only.

This should mean that adding a shim tool would be the same as adding LTI
1.1 tools currently.