EDIT deployment no longer associated with a tool
Now that we're passing the tool we want to connect to using a custom
claim, we no longer need deployment to be associated with a tool.

My mental model was wrong anyways, I thought deployment would vary based
on every instance the user adds in the LMS. Turns out that's more the
resource link id's job. Deployment might not change across the entire
LMS. But resource link id will change every time the user adds a new
LTI Shim link.

We're now storing the tool_id in the LtiSession so the PlatformLaunch
knows what tool to link to.