EDIT split LtiUser to LtiRealUser and LtiFakeUser
LtiUser holds both real and fake user info. Since we'll have a lot of
fake users for each real user, this results in a lot of data
duplication. To reduce this data duplication, I've split LtiUser into
two tables, one holding the real user info and one holding the fake user
info.

Unfortunately, this resulted in the return of the need to create the
user ahead of time on the ToolLaunch side. As I didn't want to put
login_hint into LtiSession in the future. LtiSession should be
refactored in the future to have foreign keys to tool, deployment, and
lti_real_user.

The real user info is updated every launch. It's sort of a way to ensure
that our data is kept in sync with the LMS. Not sure if this is
necessary or if we should have some other kind of sync with the LMS
instead.

Also, real users are now platform specific while fake users are tool
specific. There's no foreign key to deployment anymore as that relation
no longer makes sense.

Fixed the tests, but it's really messy now, should probably spend some
time refactoring the tests.