EDIT LtiSession add foreign key columns
LtiSession now stores foreign keys to Deployment, Tool, and LtiRealUser.
This was previously just mixed into the session JSON. The 'session' JSON
column was renamed to 'token' to make it more explicit that it should
only store the id_token payload. With the foreign keys, we can take
advantage of the built in Laravel model functions when getting the
related Deployment/Tool/LtiRealUser.

This was a suggested refactor from #5