FIX force fake email unique enforcement
For the fake email generator, it'll check if the fake email has been
used already and try to generate another one if it's been used.

Since we're planning on using some kind of email redirector, the fake
email will have to be unique for every fake user, or people will get
messages they're not supposed to get.

This is an issue brought up in #5.