ADD filter for resource link claims.
The resource link is passed in the id_token. This was previously just a
static string but is now properly being masked based on the original
resource link id.

There are other parameters that can go into a resource link claim, but
since only the id is required, I'm only passing that.