ADD course context filter, mask course id
The CourseContextFilter replaces the Canvas course id with a fake id.
Any other course information is dropped.

Still not sure if we would be better served with a single lookup table
for these ids instead of the scattered tables for each of the filters.
Trying out the independent tables for now. I'm kinda thinking that they
might not end up being simple id lookup tables, we might have to add
additional columns (masking course name, for example) later. If this
turns out not to be the case, we can always revert to a single lookup
table.