UPDATE from Laravel 6 to Laravel 7
Turns out there's a new version of Laravel already?! Luckily, there
doesn't seem to be much broken with the update.

Followed the upgrade guide to update us to Laravel 7. Looks like the
most interesting change for us is that datetime are now serialized to
ISO-8601 format, instead of the timezone-less format they used to use.
Hopefully, this increases compatibility with our Postgres timezone aware
datetime columns.