ADD Laravel autogenerated authentication system
Out of the box authentication system provided by Laravel.

Given that Laravel uses webpack, we now need to run a watcher to trigger
webpack on file change when we do UI development. UI packages will also
need to be installed as part of the setup. Readme was updated to note
this.

Honestly, this feels like we've added a lot of complications but
hopefully it actually saves time as its supposed to...

By default, Laravel expects users to self register. We don't want that,
so I've disabled it for now. Need to figure out how to convert that into
a system where admin users can add new users.

Added a default admin user to seed data:
* Email: admin@example.com
* Password: password

The way they've integrated VueJS seems like they tried to hybridized the
existing frontend system with some Vue components. It's quite odd, still
trying to get a feel for it.