ADD Laravel Passport for API authentication
Laravel's Passport package implements OAuth2 for for API authentication.
I'm enabling this for the REST API endpoints.

Unfortunately, this is a separate authentication mechanism from the
cookie based session that was generated by Laravel. So we'll probably
need to do some workarounds to get them to play nicely with each other.