EDIT enable hot module replacement
Webpack feature that allows frontend assets to be automatically updated
in the browser without you having to hit refresh. This is a separate
server that has to be run in workspace. By default, the webpack server
will refuse connections from your browser because it was configured to
only allow localhost. We need to tell the server to allow from 0.0.0.0
(any ip) so that it'll let us in.