ADD vuex, frontend API token auth system
Laravel Passport has a JSON API for obtaining Personal Access Tokens
using the session cookie authentication. This lets us obtain an API
token without having to modify the default username/password auth
system.

Passport implements OAuth2, so the token is actually a JWT.

Vuex is used for storing the API token, with a backup copy in
html5 storage.

All this requires a 2 step auth system. First the user needs to log in
and obtain the session cookie. Then we need to use that session cookie
to obtain the API token. Since the user is redirected to the home page
after logging in, I've put the token retrival call in a component on the
home page.

If the browser ever lose the token, there's no feature to regenerate it,
so the only thing we can do is revoke the old token and generate a new
one. The auth vuex module makes sure that all keys it generates are
named the same, so we only revoke keys of that name and not other keys.

Note that tokens are long lasting, 1 year by default.

Errors right now are only logged in console, should get an error system
going.