ADD user admin, only listing users work right now
The basic skeleton required for user admin, from both the front end and
the back end side.

This is going to be a bit of an odd app. As I think all our admin tools
could fit on one page for now. We can make it more complicated later.

There are places where I think vue-router would help, but I didn't want
to have two different routing systems yet. So trying to see if we can
stick with Laravel's stock routing system. Unfortunately, I suspect a
lot of the nice Vuex features might not work if we try to stick with
only Laravel routing since they were built for SPAs.

I initially wanted to use Laravel's traditional form submit based
front-end work flow, so spent some time on that. But later figured that
having a JSON API upfront might be nice later on for automation
purposes. And if we're gonna have that, might as well build the
front-end based on that anyways.