ADD vue material design icons
Added an icon set. The Material Design Icons set is huge, but Webpack
will make sure that only the icons we're using end up included in the
compiled source. But in order to do this, we'll need to import the icons
individually in our vue components. A basic example of how to do this
can be seen in the UserList or UserAdmin components.

Note that the Material Design Icons site has the icons names in
kebab-case but for import use, they need to be converted to
UpperCamelCase.