EDIT ui convert user store to store users by id
Instead of a plain array of users, we're now storing users mapped by
their id. This will make it so we don't have to loop through the entire
user array when we want to make changes to specific users.

Note that because the users list is now a javascript object, we have to
use Vue.set to add new users or Vue's reactivity will not trigger. This
is true for deleting users as well.