EDIT ui now can delete users
This will permanently delete a user from the database. Laravel has
built-in support for soft deletes that we can enable later if needed.

Delete button will disappear if there is only 1 user left. This
protection is only one UI side though, the API will happily delete the
last user if asked.

I've been wanting to have a nice "Are you sure?" button component.
Unfortunately, it's harder than expected to make one that wouldn't mess
up a mobile compatible layout. But since the admin app is so simple, I
don't think we have to worry about mobile that much, so now there's a
simple AreYouSureButton component that checks with the user to confirm
actions.

Slight renaming of user actions to remove redundancy.