EDIT UI can now edit existing users
You can only edit the name and email. Decided not to allow password
editing for now. Will need to let users change their own password later
on though.

Encountered an interesting issue with vuex while implementing this. If I
load the user info directly from userStore for editing, vuex will
complain that we're modifying the data outside of mutators. Now, there's
a few ways to fix this, but all of them involve actually allowing the
modification to go through. The problem is, I don't want the store to be
modified without a corresponding successful API call, e.g.: unless the
user save is successful, we shouldn't let the store be modified. For
some reason, the docs and posts I can find really don't mention this use
case. So I have to wonder what am I missing. For now, I'm just cloning
the user data so that we have copy that's not managed by vuex.

While I implemented the api route to get information on a specific user
and also created a call for it in userStore, it's not actually being
used, so no guarantee it's actually working.