EDIT ui now shows laravel validation errors
I've only seen the laravel validation errors, so that's the only one
that's been dealt with right now. The error notification stays in place
until dismissed, in case the user needs to reference it.

Needed to have the create/update action in userStore return a promise so
that our UserForm component doesn't navigate away if there is an error.
Otherwise, we end up with the stupid behaviour where it returns to the
user listing but the user create/edit fails, forcing you to re-enter all
the data again if you want to retry.

In order for the promise returned by the userStore to work properly, we
need to return a rejected promise in the catch block. This was folded
into the processError helper, but if we're not using that helper, we'll
need to return it manually.