ADD ui signed in user can edit their account info
Well, this became a lot more complicated than expected. Editing your own
account info was placed into a separate Laravel view, which caused
issues with users who go directly to that page being without an API
token because its the AdminMain component that signs us in.

Note that the UserController will only allow password changes for the
signed in user, they cannot change passwords for other users.

While testing this, realized that signing in from another browser
invalidates the API token held by the original browser. So I've added a
force logout when we encounter an access denied error. This is only in
the helper.processError() function though, so we'll have to use that for
all API calls that are protected by Passport.

This does mean that browser wise, you're only allowed to be signed in to
a single place. I think we can fix this later by making the API tokens
shorter lived and not bothering to manually manage them.

Replaced authStore console log errors with the notification system. Note
that the authStore cannot use helper.processError as that calls the
authStore to force logout.