FIX ui can't add new user or edit own account
When adding new user, the form fields are mistakenly disabled. The
fields are supposed to be disabled if you're editing a user and data
hasn't loaded, but shouldn't be disabled if we're adding a new user.

The SessionDropdown component wasn't loaded, so there's no way for users
to edit their own account of logout.