ADD ui extendable base vuex module, platform store
Tried to abstract out a generic vuex module that can be used as the base
for all our new vuex stores. The module base can be easily extended with
additional functions too.

There were two ways I could've done the extendable mechanism. The way I
did it was to pass in the extensions to the moduleBase function call.
The other way is in the actual implemented store, tacking on the
moduleBase's properties in the actual store. Not sure what would be best
way to do it actually yet.

Also added a platformStore to show how the generic module could be used.