ADD ui can now add/edit/delete platforms now
Added the platform API at /api/platform

Couldn't really work out how to save platform and its related data
(platform clients and keys) together all at once. So it's a multistep
save where we save the platform first and then its related data. This
might cause issues where the platform was created in the database but
the relations aren't, but api call fails, so the ui doesn't reflect
that. Will need to look into Laravel's transactional facilities to see
if they have an easy fix for this.

Also added api calls for PlatformClient and PlatformKey, but they're not
fully functional like the Platform API, only the delete was tested. The
others will return a 501 Not Implemented error. The paths are
conventional:
* /api/platform/{platformId}/clients/{clientId}
* /api/platform/{platformId}/keys/{clientId}

For the platform's clients and keys, the only two operations we have are
add and delete. Resulting in their respective APIs also being limited
too.

The platformStore now also shows how to extend the moduleBase, with
additional actions for platform's clients and keys.

I tried a few JWT libraries to see if they can do a proper validation of
the JWKs being entered for the platform's keys, but they all had issues
or were of limited use. So we ended up with a barebones validation that
only makes sure that we have all the necessary fields in the JWK. It
doesn't check to see if it's a proper RSA key, that'll only show up when
you actually try to do a LTI launch.

Note that we're allowing a mix of automatically retrieved keys from the
JWKS URL and manually entered JWK. I didn't make the two mechanisms
exclusive from each other. Users can enter a JWKS URL *and* enter some
keys manually if they want.

The component event system is more complicated than I'd like. E.g.: As
delete API calls have to propagate up to parent components. I'm not sure
how to resolve this.