ADD ui can now add/edit/delete LTI tools
Simpler than the platform since there's no clients. A lot of the code is
similar, so I know there's an opportunity to abstract them out into
generic components, but I'm not sure how best to do that yet.

Unlike for the platform, the public keys aren't required for the tool
for LTI launches. They're only required for LTI service calls where the
tool talks to the platform directly.

There's also no unique requirements on the fields, but now thinking
about it, we should enforce uniqueness on client_id. So I've made tools
client_id a unique field.

Generalized the JWK section to a JwkForm component. Since there's two
sets of values that the JwkForm component needs to pass, we can't use
v-model. So we have to manually do what v-model does under the hood for
the two values we need to pass, the jwks url and the jwk keys. There's
also a deleteJwk event that needs to bubble up.